Get rate of consumption of resource by fish

This function is used by the dynamics functions, for example detritus_dynamics() and carrion_dynamics(), in their calculation of the resource biomasses at the next time step. Not exported.

getConsumptionByFish(params, resource, n, rates)

Arguments

params
The MizerParams object

resource
A string with the name of the resource for which the consumption is to be calculated

n
A matrix with the current size spectrum of fish

rates
A list with the rates

Value

A number giving the rate at which the resource is consumed by all the fish in units of grams/year