Get encounter rate for unstructured resource

Calculates the rate \(E_i(w)\) at which a predator of species \(i\) and weight \(w\) encounters resources (grams/year).

getUREncounter(
  params,
  n = params@initial_n,
  n_pp = params@initial_n_pp,
  n_other = params@initial_n_other,
  ...
)

Arguments

params
A MizerParams object

n
A matrix of species abundances (species x size)

n_pp
A vector of the plankton abundance by size

n_other
A list of abundances of other dynamical components

Value

A two dimensional array (predator species x predator size)

Resource encounter

In addition to the contribution from predation on fish prey and plankton, the food encounter rate may have a contribution from unstructured resource components. This takes the form $$E_{u.i} = \sum_d \rho_{id}(w) B_d.$$ where \(B_d\) is the biomass of the d-th unstructured resource component and \(\rho_{id}(w)\) is a parameter that therefore determines the rate at which a predator of species \(i\) and size \(w\) encounters biomass from the d-th unstructured resource component. This is set with setResourceEncounter.

The encounter rate is multiplied by \(1-f_0\) to obtain the consumption rate, where \(f_0\) is the feeding level calculated with getFeedingLevel. This is used by the project function for performing simulations.

The function returns values also for sizes outside the size-range of the species. These values should not be used, as they are meaningless.

Examples

if (FALSE) {
# TODO
}